requirements.txt
Bootstrap_Flask==2.2.0
Flask==2.3.2
python-dotenv==1.0.1
PyPDF2==3.0.1
